Fix time

+++ models/sphinx.go
@@ -28,9 +28,9 @@ func SphinxSearch(keyword string, page, pageSize int) (interface{}, error) {
 			postIntro := strings.Replace(intro, strings.ToUpper(keyword), "<b style='color:red'>"+strings.ToUpper(keyword)+"</b>", -1)
 			tempData["post_title"] = strings.Replace(postIntro, strings.ToLower(keyword), "<b style='color:red'>"+strings.ToLower(keyword)+"</b>", -1)
 		}
-		postAt, ok := match.AttrValues[2].(int64)
+		postAt, ok := match.AttrValues[2].(uint32)
 		if ok {
-			postDate := strings.Split(time.Unix(postAt, 0).Format("2006-01-02"), " ")
+			postDate := strings.Split(time.Unix(int64(postAt), 0).Format("2006-01-02"), " ")
 			tempData["post_date"] = postDate[0]
 		} else {
 			tempData["post_date"] = ""
